Create Exercise Tracker
This Exercise tracker allows the user to enter a natural english language sentence describing their exercise for the day/ 

Example: 'Today I walked for 10 minutes and lifted weights for an hour'. 

Using a Nutronix end point, the sentence is parsed and the output is sent and saved to a google sheet using a Sheety endpoint. 
This means that the final result is that the sentence becomes the basis for table of data that shoes the activity, length and calories burnt.